move lriter
