+++ python/src/lib.rs
@@ -6,11 +6,11 @@ use pyo3::{
 
 use lrcalc::helper::{
     ivlincomb::LinearCombination,
-    lriter::{lrit_good, lrit_new, lrit_next},
     part::{part_qdegree, part_qentry},
     schublib::{mult_schubert, mult_schubert_str, trans},
     schur::{schur_coprod, schur_lrcoef, schur_mult, schur_mult_fusion, schur_skew},
 };
+use lrcalc::lriter::{lrit_good, lrit_new, lrit_next};
 
 pub fn ivlc_to_dict_of_vecs(lc: LinearCombination) -> Vec<(Vec<i32>, i32)> {
     let mut ans = Vec::new();
