+++ src/lrcalc/helper/mod.rs
@@ -1,7 +1,6 @@
 pub mod ivector;
 pub mod ivlincomb;
 pub mod lrcoef;
-pub mod lriter;
 pub mod optim;
 pub mod part;
 pub mod perm;
