+++ src/lrcalc/helper/schur.rs
@@ -1,10 +1,10 @@
+use super::super::lriter::{lrit_expand, lrit_good, lrit_new, lrit_next, LRTableauIterator};
 use super::ivector::{iv_hash, IntVector};
 use super::ivlincomb::{
     ivlc_add_element, ivlc_new, ivlc_new_default, LinearCombination, LC_COPY_KEY, LC_FREE_KEY,
     LC_FREE_ZERO,
 };
 use super::lrcoef::lrcoef_count;
-use super::lriter::{lrit_expand, lrit_good, lrit_new, lrit_next, LRTableauIterator};
 use super::optim::{optim_coef, optim_fusion, optim_mult, optim_skew};
 use super::part::{part_entry, part_valid};
